Renamed instructions to components in Adapter
