+++ stwo_cairo_prover/crates/adapted_prover/src/main.rs
@@ -63,8 +63,8 @@ fn run(args: impl Iterator<Item = String>) -> Result<CairoProof<Blake2sMerkleHas
     let vm_output: CairoInput =
         import_from_vm_output(args.pub_json.as_path(), args.priv_json.as_path())?;
 
-    let instruction_counts = vm_output.instructions.counts();
-    log::info!("Instruction counts: {instruction_counts:?}");
+    let component_counts = vm_output.components_usage.counts();
+    log::info!("Component counts: {component_counts:?}");
 
     let proof = prove_cairo(vm_output)?;
